Create embedded-game-controller library
Move code around with just the minimum amount of changes to keep it
compiling and working. More cleanups will happen later.

Add a client.h header, where we'll put all the client API of the
library. Note that currently the library has some undefined
dependencies, which are resolved only when linking with the fakemote
client. This will be fixed in a later commit.

f